+++ CHANGELOG.md
@@ -59,6 +59,7 @@
 
 - FEAT: calling class constructors
 - FEAT: accessing public class fields
+- FEAT: accessing public class methods
 
 ## TODO